feat: useTiptap 훅 생성 후 로직 분리
