+++ client/src/components/write/Tiptap/index.tsx
@@ -1,32 +1,13 @@
 import React, { useEffect } from 'react';
-import { useEditor, EditorContent } from '@tiptap/react';
+import { EditorContent, Editor } from '@tiptap/react';
 import ToolBar from '../Toolbar';
-
-//tiptap
-import StarterKit from '@tiptap/starter-kit';
-import Highlight from '@tiptap/extension-highlight';
-import Image from '@tiptap/extension-image';
-import CodeBlockLowlight from '@tiptap/extension-code-block-lowlight';
-
-import { common, createLowlight } from 'lowlight';
 
 interface TiptapProps {
   content: string;
+  editor: Editor | null;
 }
 
-const Tiptap = ({ content }: TiptapProps) => {
-  const lowlight = createLowlight(common);
-  const editor = useEditor({
-    extensions: [
-      StarterKit,
-      Highlight,
-      Image.configure({ inline: true, allowBase64: true }),
-      CodeBlockLowlight.configure({
-        lowlight,
-      }),
-    ],
-  });
-
+const Tiptap = ({ content, editor }: TiptapProps) => {
   useEffect(() => {
     if (content) {
       editor?.commands.setContent(content);
